deprecated watchd

+++ maica/rotate_cache.py
@@ -1,60 +1,28 @@
-import nest_asyncio
-nest_asyncio.apply()
 import asyncio
-import aiomysql
 import datetime
+import schedule
+
+from typing import *
 from maica_utils import *
 
-class rotation_instance():
-
-    def __init__(
-        self,
-        host = load_env('DB_ADDR'), 
-        user = load_env('DB_USER'),
-        password = load_env('DB_PASSWORD'),
-        authdb = load_env('AUTH_DB'),
-        maicadb = load_env('MAICA_DB'),
-        test = False
-    ):
-        self.host, self.user, self.password, self.authdb, self.maicadb, self.test = host, user, password, authdb, maicadb, test
-        asyncio.run(self.create_pool())
-        asyncio.run(self.rotate_caches())
-    
-    async def create_pool(self):
-        self.loop = asyncio.get_event_loop()
-        self.maicapool = await aiomysql.create_pool(host=self.host, user=self.user, password=self.password, loop=self.loop, db=self.maicadb, autocommit=True)
-    
-    async def send_modify(self, expression, values=None):
-        pool = self.maicapool
-        async with pool.acquire() as conn:
-            async with conn.cursor() as cur:
-                if not values:
-                    await cur.execute(expression)
-                else:
-                    await cur.execute(expression, values)
-                await conn.commit()
-                lrid = cur.lastrowid
-        return lrid
-    
-    async def send_query(self, expression, values=None, fetchall=False):
-        pool = self.maicapool
-        async with pool.acquire() as conn:
-            async with conn.cursor() as cur:
-                if not values:
-                    await cur.execute(expression)
-                else:
-                    await cur.execute(expression, values)
-                results = await cur.fetchone() if not fetchall else await cur.fetchall()
-        return results
-
-    async def rotate_caches(self, keep_time=load_env('ROTATE_MSCACHE')):
+async def rotate_cache(maica_pool: Optional[DbPoolCoroutine]=None):
+    maica_pool = default(maica_pool, ConnUtils.maica_pool())
+    keep_time = load_env('ROTATE_MSCACHE')
+    if int(keep_time):
         timestamp = datetime.datetime.now()
         sql_expression_1 = "SELECT spire_id, timestamp FROM ms_cache"
         sql_expression_2 = "DELETE FROM ms_cache WHERE spire_id = %s"
-        result = await self.send_query(sql_expression_1, None, True)
+        result = await maica_pool.query_get(sql_expression_1)
         for row in result:
             if row[1] + datetime.timedelta(hours=int(keep_time)) <= timestamp:
-                await self.send_modify(sql_expression_2, (row[0]))
+                await maica_pool.query_modify(sql_expression_2, (row[0]))
 
-if __name__ == '__main__':
-    r = rotation_instance()
+def wrap_rotate_cache(*args, **kwargs):
+    asyncio.run(rotate_cache(*args, **kwargs))
+
+async def schedule_rotate_cache(*args, **kwargs):
+    if load_env('ROTATE_MSCACHE') != '0':
+        schedule.every().day.at("04:00").do(wrap_rotate_cache, *args, **kwargs)
+    while True:
+        schedule.run_pending()
+        asyncio.sleep(60)
